break down api/filters.go separate file for dimensions
